refact(const):rename const package name
