+++ pkg/constant/url.go
@@ -1,4 +1,4 @@
-package _default
+package constant
 
 const (
 	GatewayDeployScripts = "https://github.com/aenjoy/iot-lubricant/raw/refs/heads/main/scripts/gateway_init.sh"
