+++ pkg/model/agent.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/aenjoy/iot-lubricant/pkg/default"
+	"github.com/aenjoy/iot-lubricant/pkg/constant"
 	"github.com/aenjoy/iot-lubricant/pkg/logger"
 	"github.com/aenjoy/iot-lubricant/pkg/types/container"
 	"github.com/aenjoy/iot-lubricant/pkg/types/operation"
@@ -138,7 +138,7 @@ var AgentContainer = container.Container{
 	Name:    "lubricant-agent",
 	Network: network.NetworkBridge,
 	ExposePort: map[string]int{
-		fmt.Sprintf("%d", _default.AgentGrpcPort): 0,
+		fmt.Sprintf("%d", constant.AgentGrpcPort): 0,
 	},
 }
