+++ pkg/ssh/client.go
@@ -8,7 +8,7 @@ import (
 	"net/http"
 	"os"
 
-	def "github.com/aenjoy/iot-lubricant/pkg/default"
+	def "github.com/aenjoy/iot-lubricant/pkg/constant"
 	"github.com/aenjoy/iot-lubricant/pkg/logger"
 	"github.com/aenjoy/iot-lubricant/pkg/model"
 	"github.com/pkg/sftp"
