+++ services/logg/app.go
@@ -1,1 +1,4 @@
 package logg
+
+type app struct {
+}
