+++ services/lubricant/casdoor.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"os"
 
-	def "github.com/aenjoy/iot-lubricant/pkg/default"
+	def "github.com/aenjoy/iot-lubricant/pkg/constant"
 	"github.com/aenjoy/iot-lubricant/pkg/logger"
 	"github.com/aenjoy/iot-lubricant/services/lubricant/config"
 	"github.com/casdoor/casdoor-go-sdk/casdoorsdk"
