+++ services/lubricant/key.go
@@ -3,7 +3,7 @@ package lubricant
 import (
 	"os"
 
-	def "github.com/aenjoy/iot-lubricant/pkg/default"
+	def "github.com/aenjoy/iot-lubricant/pkg/constant"
 	"github.com/aenjoy/iot-lubricant/pkg/logger"
 	"github.com/aenjoy/iot-lubricant/pkg/utils/file"
 	"github.com/aenjoy/iot-lubricant/pkg/utils/hash"
